refactor: move components into pages
